Replace single quotes with double quotes
Preferences change, preferably towards the beginning of a project.

Changes were made either where the contained text is user feedback or where the text is SQL. For user feedback, this is so that it's easier to change the feedback later, if need be, to something that contains apostrophes. For SQL text, this is so the text can contain string values (surrounded by single quotes).